spec/pyproject.toml
[project]
name = "task-management-assistant"
version = "0.1.0"
description = "AI-powered task management assistant using Strands and OpenRouter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents[litellm]>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_assistant"]
